Bump `build/number` to `2`
While the package should be largely unchanged, go ahead and bump the
`build/number` for good measure. This will also help indicate that the
CUDA versions supported has changed. If issues do arise (though not
expected), this will make it easier to identify and pull the package
(should it be needed), but again not expecting any need.