Drop `yum_requirements.txt`
As `libnuma` is already handled as a dependency, there should be no need
to grab it from the system any more. As this is the only thing in
`yum_requirements.txt`, go ahead and drop that file altogether. This
should also make it clearer to users that no system packages should be
needed to support `ucx` at this point.